Create newfunc%forloop
This program is an improvement of the previous factorial and summation program.

Now, we don't need to define the value o n for each function. The user just need to assign a value of n and this value will work for all functions. 

To summarize this, let's suppose that the user needs to find the factorial and the sum of the first 5 integers. The previous program asked the user for 5 two times: one for the factorial and the other for the sum. With this program, the user just need to input 5 once.